% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stipendrep.R
\name{stipend_new}
\alias{stipend_new}
\title{function to calculate new stipend and creat sheets}
\usage{
stipend_new(new_url, month)
}
\arguments{
\item{new_url}{input url}

\item{month}{input month in Month_Year format}
}
\description{
function to calculate new stipend and creat sheets
}
