% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stipendrep.R
\name{stipend_old}
\alias{stipend_old}
\title{Stipend Calculation and sheet generation}
\usage{
stipend_old(url, month_year)
}
\arguments{
\item{url}{input url of the the respective stipend sheet}

\item{month_year}{input month in Month_year format}
}
\description{
Stipend Calculation and sheet generation
}
